backend/requirements.txt
fastapi>=0.115.0
uvicorn
sqlalchemy[asyncio]
aiosqlite
pydantic
pydantic-settings
python-multipart
python-dotenv
pytest
httpx
langchain
langchain-core
sentence-transformers
faiss-cpu
langchain-community
langchain-huggingface
mcp>=1.25.0
anyio>=4.5.0
starlette>=0.49.0
langchain-google-genai
sse-starlette
